Made abort a real plugin and fixed tests
