+++ flexget/tests/notifiers/test_notify_crash.py
@@ -3,20 +3,6 @@ from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 
 import mock
 
-from flexget import plugin
-from flexget.event import event
-
-
-class AbortPlugin(object):
-    def on_task_output(self, task, config):
-        task.abort('abort plugin')
-
-
-@event('plugin.register')
-def register():
-    plugin.register(AbortPlugin, 'abort', debug=True, api_ver=2)
-
-
 @mock.patch('requests.Session.request')
 class TestNotifyCrash(object):
     config = """
@@ -30,24 +16,26 @@ class TestNotifyCrash(object):
 
             notify_crash:
               to:
-                - pushbullet:
-                    apikey: "apikey"
+                - pushover:
+                    token: token
+                    user_key: user_key
           no_crash:
             disable: builtins
             notify_crash:
               to:
-                - pushbullet:
-                    apikey: "apikey"
+                - pushover:
+                    token: token
+                    user_key: user_key
     """
 
     def test_abort(self, mocked_request, execute_task):
         execute_task('test_crash', abort=True)
-        data = {'body': 'Task returned 0 accepted entries', 'title': 'test_crash'}
+        data = {'message': 'Task returned 0 accepted entries', 'title': 'test_crash'}
 
         assert mocked_request.called
 
         for k, v in data.items():
-            assert mocked_request.call_args[1]['json'][k] == v
+            assert mocked_request.call_args[1]['data'][k] == v
 
     def test_no_crash(self, mocked_request, execute_task):
         execute_task('no_crash')
